export interpolateQueryExpr for metrics drilldown links
