+++ packages/grafana-prometheus/src/index.ts
@@ -58,6 +58,7 @@ export { PrometheusDatasource } from './datasource';
 // The parts
 export { addLabelToQuery } from './add_label_to_query';
 export { type QueryEditorMode, type PromQueryFormat, type Prometheus } from './dataquery';
+export { interpolateQueryExpr } from './escaping';
 export { loadResources } from './loadResources';
 export { PrometheusMetricFindQuery } from './metric_find_query';
 export { promqlGrammar } from './promql';
